#108: Raster data (http://github.com/Polymap4/polymap4-p4/issues/issue/108)
AllResolver: cache CompletableFuture instead of the resulting IResolvableInfo.

Working on this started when deadlocks occured when using raster sources. Seemed that using a Cache and doing the resolve inside the get wasn't a good idea because resolving might block the entire cache and held a lock on it. So it seemed to be much better to have a Future cached and just block the threads that are interested in the same resolvable. Later this caused the IMetadataContentResolver API to be changed accordingly.